+++ FrogHouse/FrogHouse/Application/AppDelegate.swift
@@ -12,10 +12,18 @@ import FirebaseCore
 class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         FirebaseApp.configure()
-        do {
-            try insertMockVideos()
-        } catch {
-            
+//        do {
+//            try insertMockVideos()
+//        } catch {
+//            
+//        }
+//        return true
+        Task {
+            do {
+                try await FirestoreVideoListStore.shared.loadFirestoreData()
+            } catch {
+                print("🔥 초기 데이터 로딩 실패: \(error)")
+            }
         }
         return true
     }
