+++ FrogHouse/FrogHouse/Source/Scene/RecommendedVideo/RecommendedVideoViewModel.swift
@@ -29,33 +29,35 @@ final class RecommendedVideoViewModel {
     // MARK: Jay - 초기화
     init(store: FirestoreVideoListStore = .shared) {
         self.store = store
-        // MARK: Jay - 필요 시 실시간 리스닝 시작 (원치 않으면 주석)
-        // store.startListening(orderBy: "viewCount", descending: true)
+        // MARK: Jay - 필요 시 실시간 리스닝 (원치 않으면 주석 처리)
+        // store.startListening()
     }
 
-    // MARK: Jay - 로드
+    // MARK: Jay - 로드 (캐시 우선, TTL 지나면 자동 갱신)
     func load() {
         Task {
             await MainActor.run { onLoadingChanged?(true) }
             do {
-                // MARK: Jay - DTO 가져오기 (TTL 고려)
-                let dtos = try await store.getOrFetch(orderBy: "viewCount", descending: true)
+                // MARK: Jay - DTO 가져오기 (캐시 유효하면 캐시, 아니면 Firestore에서 갱신)
+                let dtos = try await store.getOrFetch()
 
-                // MARK: Jay - 정렬 (recent → viewCount) 원하면 바꾸기
+                // MARK: Jay - 정렬 (lastWatchedAt 우선 → viewCount)
                 let sorted = dtos.sorted(by: FirestoreVideoListMapper.sortByRecentThenViewCount)
 
-                // MARK: Jay - 매핑 후 상위 N개
-                var models: [RecommendedVideoItem] = []
-                models.reserveCapacity(maxCount)
+                // MARK: Jay - 매핑 후 상위 N개만 사용
+                var tmp: [RecommendedVideoItem] = []
+                tmp.reserveCapacity(maxCount)
                 for dto in sorted {
                     if let m = FirestoreVideoListMapper.toRecommendedVideoItem(dto) {
-                        models.append(m)
-                        if models.count >= maxCount { break }
+                        tmp.append(m)
+                        if tmp.count >= maxCount { break }
                     }
                 }
 
-                await MainActor.run {
-                    self.items = models
+                // MARK: Jay - Swift 6 동시성: 불변 스냅샷으로 UI 업데이트
+                let finalModels = tmp
+                await MainActor.run { [finalModels] in
+                    self.items = finalModels
                     self.currentIndex = min(self.currentIndex, max(0, self.items.count - 1))
                     self.onListUpdated?(self.items)
                     self.onLoadingChanged?(false)
@@ -63,6 +65,7 @@ final class RecommendedVideoViewModel {
                         self.onCurrentItemChanged?(self.items[self.currentIndex], self.currentIndex, self.items.count)
                     }
                 }
+
             } catch {
                 await MainActor.run {
                     self.onLoadingChanged?(false)
